+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Log solaredge API request and response for debugging.
+
 ### Changed
 - Reverted needless logging and error handling changes.